pyproject.toml
[tool.poetry]
name = "matosinhosgrocery"
version = "0.1.0"
description = "Grocery list and receipt management bot for Matosinhos."
authors = ["woodwowood@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.0"
python-telegram-bot = "^21.1.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
aiosqlite = "^0.20.0"
openai = "^1.28.0"
google-api-python-client = "^2.130.0"
google-auth-oauthlib = "^1.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.6.0"
ruff = "^0.1.0" # For linting

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 